# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Marketplace API",
        "description": "Use this API to manage product data.",
        "contact": {
            "email": "fathir1234@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "ProductsAPI"
        }
    ],
    "paths": {
        "/api/brands": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "Get all brands",
                "description": "Retrieve a list of all brands, with optional filtering by keyword",
                "operationId": "3c41aa2e65314941e37a634e810c5f73",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "A keyword to filter brands by their name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of brands",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "brands": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Brand"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Brands"
                ],
                "summary": "Create a new brand",
                "description": "Create a new brand with name, address, and email",
                "operationId": "1dae5363a453b159e035bc04b3fbc0be",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "brand_name",
                                    "brand_address",
                                    "brand_email"
                                ],
                                "properties": {
                                    "brand_name": {
                                        "type": "string",
                                        "example": "Acme Corporation"
                                    },
                                    "brand_address": {
                                        "type": "string",
                                        "example": "123 Main St, Springfield"
                                    },
                                    "brand_email": {
                                        "type": "string",
                                        "example": "info@acme.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Brand created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Brand"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/brands/{id}": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "Get a specific brand",
                "description": "Retrieve a specific brand by its ID",
                "operationId": "767f67c0955810b713c48d4100f02362",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the brand",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Brand retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Brand"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Brand not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Brand not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Brands"
                ],
                "summary": "Update a specific brand",
                "description": "Update the name, address, and/or email of a specific brand by its ID",
                "operationId": "1fac61ef4f1f3db38169272ad0e5ed09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the brand",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "brand_name",
                                    "brand_address",
                                    "brand_email"
                                ],
                                "properties": {
                                    "brand_name": {
                                        "type": "string",
                                        "example": "Acme Corporation"
                                    },
                                    "brand_address": {
                                        "type": "string",
                                        "example": "123 Main St, Springfield"
                                    },
                                    "brand_email": {
                                        "type": "string",
                                        "example": "info@acme.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Brand updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Brand"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Brand not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Brand not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Brands"
                ],
                "summary": "Delete a specific brand",
                "description": "Delete a specific brand by its ID",
                "operationId": "401903a3c5a05c699953cb0b07a385e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the brand",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Brand deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Brand deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Brand not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Brand not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Brand": {
                "required": [
                    "brand_name",
                    "brand_address",
                    "brand_email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "brand_name": {
                        "type": "string",
                        "example": "Acme Corporation"
                    },
                    "brand_address": {
                        "type": "string",
                        "example": "123 Main St, Springfield"
                    },
                    "brand_email": {
                        "type": "string",
                        "example": "info@acme.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-26T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-26T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter your bearer token in the format **Bearer &lt;token>**",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}